requirements.txt
matplotlib~=3.2.2
numpy~=1.17.0
pandas~=1.1.3
torchvision~=0.8.2
tensorboardx~=2.4